Updated participle, supported modules, and minor science changes
Too much for a single PR in the normal case. :-(

Looks like ScienceTerm got changed to ScienceSubject, with struct 
changes to match.

Minor changes to main to list the first ten science subjects.

Lexer code was updated to match current participle logic.

Finally, go modules are now supported, and the .gitignore was updated.